+++ NEWS.md
@@ -4,8 +4,9 @@ Development version
 
 - `learner_surv_cox`, `learner_surv_rf`, `cumhaz`
 - feature(superlearner): more robust metalearner (in case quadratic prog. fails)
-- feature(learner_cox): mets::phreg wrapper. Also disabling openMP in
-  RcppArmadillo
+- feature(learner_cox): mets::phreg wrapper.
+- disabling openMP in RcppArmadillo
+- bugfix: issue with `cate` with repeated cross-fitting fixed.
 
 # targeted 0.7.1